Fix errors, typos, and improve test coverage
This commit addresses several issues in the codebase, including:

- **Critical Bug Fix:** Corrected the iOS platform version in `Package.swift` from `.v26` to a valid version, `.v16`.
- **Refactoring:**
    - Refactored `Camera.swift` to improve code quality, fix bugs, and ensure consistent data types.
    - Refactored `CameraModel.swift` to correctly handle state updates and improve robustness.
    - Refactored `CameraView.swift` to enhance the user experience.
- **Test Coverage:** Significantly improved test coverage by adding tests for photo capture, camera controls, and state transitions in `CameraModelTests.swift`.
- **Documentation:** Updated the `README.md` file to reflect API changes and added a section on how to run the tests.
- **Code Review Feedback:** Addressed feedback from the code review, including fixing a failing test case.

+++ Sources/Camera/CameraProtocol.swift
@@ -23,6 +23,6 @@ protocol CameraProtocol: Actor {
     func takePhoto() async
     func changeCodec(_ codec: VideoCodecType)
     func changeFlashMode(_ flashMode: CameraFlashMode)
-    func changeZoom(_ factor: Float) throws
+    func changeZoom(_ factor: CGFloat) throws
     func changePosition() async throws
 }
